Fix build.

+++ OpenSim/Region/CoreModules/World/Terrain/PaintBrushes/ErodeSphere.cs
@@ -160,8 +160,8 @@ namespace OpenSim.Region.CoreModules.World.Terrain.PaintBrushes
             // Using one 'rain' round for this, so skipping a useless loop
             // Will need to adapt back in for the Flood brush
 
-            ITerrainChannel water = new TerrainChannel(BrushSize, BrushSize, null);
-            ITerrainChannel sediment = new TerrainChannel(BrushSize, BrushSize, null);
+            ITerrainChannel water = new TerrainChannel((int)BrushSize, (int)BrushSize, null);
+            ITerrainChannel sediment = new TerrainChannel((int)BrushSize, (int)BrushSize, null);
 
             // Fill with rain
             for (x = 0; x < BrushSize; x++)
@@ -175,11 +175,11 @@ namespace OpenSim.Region.CoreModules.World.Terrain.PaintBrushes
                 {
                     for (y = 0; y < water.Height; y++)
                     {
-                        if (mask[rx+x, ry+y])
+                        if (mask[(int)rx + x, (int)ry + y])
                         {
                             const double solConst = (1.0 / rounds);
                             double sedDelta = water[x, y] * solConst;
-                            map[rx+x, ry+y] -= sedDelta;
+                            map[(int)rx + x, (int)ry + y] -= sedDelta;
                             sediment[x, y] += sedDelta;
                         }
                     }
@@ -197,7 +197,7 @@ namespace OpenSim.Region.CoreModules.World.Terrain.PaintBrushes
 
                         int neighbours = 0;
                         double altitudeTotal = 0.0;
-                        double altitudeMe = map[rx+x, ry+y] + water[x, y];
+                        double altitudeMe = map[(int)rx + x, (int)ry + y] + water[x, y];
 
                         const int NEIGHBOUR_ME = 4;
                         const int NEIGHBOUR_MAX = 9;
@@ -297,10 +297,10 @@ namespace OpenSim.Region.CoreModules.World.Terrain.PaintBrushes
                         double sedimentDeposit = sediment[x, y] - waterCapacity;
                         if (sedimentDeposit > 0)
                         {
-                            if (mask[rx+x, ry+y])
+                            if (mask[(int)rx + x, (int)ry + y])
                             {
                                 sediment[x, y] -= sedimentDeposit;
-                                map[rx+x, ry+y] += sedimentDeposit;
+                                map[(int)rx + x, (int)ry + y] += sedimentDeposit;
                             }
                         }
                     }
@@ -310,8 +310,8 @@ namespace OpenSim.Region.CoreModules.World.Terrain.PaintBrushes
             // Deposit any remainder (should be minimal)
             for (x = 0; x < water.Width; x++)
                 for (y = 0; y < water.Height; y++)
-                    if (mask[rx+x, ry+y] && sediment[x, y] > 0)
-                        map[rx+x, ry+y] += sediment[x, y];
+                    if (mask[(int)rx + x, (int)ry + y] && sediment[x, y] > 0)
+                        map[(int)rx + x, (int)ry + y] += sediment[x, y];
         }
 
         #endregion
